Better rate limit

+++ tools/python-data/data/rate_limit.php
@@ -1,31 +1,39 @@
 <?php
 
 function check_rate_limit($database, $ipaddr, $q) {
-    $db = new \SQLite3($database);
-    $db->exec("CREATE TABLE IF NOT EXISTS Access (ipaddr TEXT, address TEXT, retrieved_at DATETIME DEFAULT CURRENT_TIMESTAMP)");
-    $db->exec("DELETE FROM Access WHERE retrieved_at <= date('now', '-1 day')");
 
+    if ( ! is_string($ipaddr) ) return 1;
+    $db = false;
+    try {
+        $db = new \SQLite3($database);
+        $db->enableExceptions(true);
+        $db->exec("CREATE TABLE IF NOT EXISTS Access (ipaddr TEXT UNIQUE, retrieved_at DATETIME DEFAULT CURRENT_TIMESTAMP)");
+        $db->exec("DELETE FROM Access WHERE retrieved_at <= date('now', '-1 day')");
 
-    $stm = $db->prepare("SELECT address, julianday('now'), julianday(retrieved_at) FROM Access WHERE ipaddr = :ip ORDER BY retrieved_at DESC LIMIT 1");
-    $stm->bindValue(':ip', $ipaddr, SQLITE3_TEXT);
+        $stm = $db->prepare("SELECT julianday('now'), julianday(retrieved_at) FROM Access WHERE ipaddr = :ip ORDER BY retrieved_at DESC LIMIT 1");
+        $stm->bindValue(':ip', $ipaddr, SQLITE3_TEXT);
 
-    $res = $stm->execute();
-    $row = $res->fetchArray(SQLITE3_NUM);
+        $res = $stm->execute();
+        $row = $res->fetchArray(SQLITE3_NUM);
 
-    $delta = 1000;
-    if ( $row ) {
-        $now = $row[1];
-        $retrieved_at = $row[2];
-        $delta = intval(($now - $retrieved_at) * (24*60*60));
+        $delta = 1000;
+        if ( $row ) {
+            $now = $row[0];
+            $retrieved_at = $row[1];
+            $delta = intval(($now - $retrieved_at) * (24*60*60));
+        }
+
+        $stm = $db->prepare("INSERT OR REPLACE INTO Access (ipaddr, retrieved_at) VALUES (:ip, datetime('now'));");
+        $stm->bindValue(':ip', $ipaddr, SQLITE3_TEXT);
+
+        $res = $stm->execute();
+
+    } catch(\Exception $e) {
+       error_log("check_rate_limit error ".$e->getMessage());
+       $delta = 1;
     }
 
-    $stm = $db->prepare('INSERT INTO Access (ipaddr, address) VALUES (:ip, :address);');
-    $stm->bindValue(':ip', $ipaddr, SQLITE3_TEXT);
-    $stm->bindValue(':address', $q, SQLITE3_TEXT);
-
-    $res = $stm->execute();
-
-    $db->close();
+    if ( $db ) $db->close();
 
     return $delta;
 }
